complete project

+++ tour/program/views.py
@@ -8,7 +8,7 @@ from program.models import Program, Picture
 from program.forms import *
 
 
-def create(request):
+def create_program(request):
     if not request.user.is_authenticated():
         return redirect('/login')
 
@@ -22,33 +22,32 @@ def create(request):
             program = Program(title=title, description=description, price=price)
             program.save()
 
-            url = "xxxxxa"
             #url = form.data.get('picture')
             #picture = Picture(url=url, related_program=program)
             #picture.save()
 
-            return render_to_response('view_program.html', {'program': program, 'url':url})
+            url = request.FILES['picture']
+            picture = Picture(url=url, related_program=program)
+            picture.save()
+
+            return redirect('view_program',program.id)
 
     else:
         form = CreateForm()
-
-    #context = {
-    #    'action': '',
-    #    'form': form,
-    #}
-    #context.update(csrf(request))
-
-    #return render_to_response('create.html', context)
-    
     context = RequestContext(request, {
         'action': '',
         'form': form,
     })
 
-    return render_to_response('create.html', context)
+    return render_to_response('create_program.html', context)
 
 def edit_program(request, program_id):
+    if not request.user.is_authenticated():
+        return redirect('/login')
+
     program = Program.objects.get(id=program_id)
+    picture = Picture.objects.get(related_program = program)
+    url = picture.url
     if request.method == 'POST':
         form = CreateForm(request.POST, request.FILES)
 
@@ -59,53 +58,65 @@ def edit_program(request, program_id):
 
             program.save()
 
-            url = "xxxxx"
-            #url = form.data.get('picture')
-            #picture = Picture(url=url, related_program=program)
-            #picture.save()
+            if request.FILES:
+                picture.url = request.FILES['picture']
+                picture.save()
 
-            return render_to_response('view_program.html', {'program': program, 'url':url})
+            return redirect('view_program', program_id)
 
     else:
         defaults = {
             'title': program.title,
             'description': program.description,
             'price': program.price,
+            'picture': url,
 
         }
         form = CreateForm(defaults)
         
     context = RequestContext(request, {
-        'action': '',
         'form': form,
+        'picture': picture,
     })
 
     return render_to_response('edit_program.html', context)
 
 def delete_program(request, program_id):
+    if not request.user.is_authenticated():
+        return redirect('/login')
+
     program = Program.objects.get(id=program_id)
     program.delete()
 
     program_all = Program.objects.all()
     return redirect('/', {'programs': program_all})
 
+def manage_program(request):
+    if not request.user.is_authenticated():
+        return redirect('/login')
 
+    programs = Program.objects.all()
+    context = RequestContext(request, {
+        'programs': programs,
+    })
 
-def view_all(request):
+    return render_to_response('manage_program.html', context)
+
+def view_program_all(request):
     programs = Program.objects.all()
 
     context = RequestContext(request, {
         'programs': programs,
     })
 
-    return render_to_response('view_all.html', context)
+    return render_to_response('view_program_all.html', context)
 
 def view_program(request, program_id):
     program = Program.objects.get(id=program_id)
-    url = 'image is temporary disabled ^^" '
+    picture = Picture.objects.get(related_program = program)
     context = RequestContext(request, {
         'program': program,
-        'url': url,
+        'picture': picture,
     })
 
     return render_to_response('view_program.html', context)
